examples/agent-integration/files/s3-translation-presigned/pyproject.toml
[project]
name = "s3-translation-presigned"
version = "0.1.0"
description = "Example demonstrating S3 extension for file upload/download via pre-signed URLs using a mock translation agent."
authors = [
    { name = "IBM Corp." },
]
requires-python = ">=3.11,<3.14"
dependencies = [
    "agentstack-sdk[s3]>=0.6.3",
    "asyncclick>=8.1.8",
    "python-dotenv>=1.2.1",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.uv.sources]
agentstack-sdk = { path = "../../../../apps/agentstack-sdk-py", editable = true }

[project.scripts]
server = "s3_translation_presigned.agent:run"
client = "s3_translation_presigned.client:run"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "watchfiles>=1.1.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
